Removed hasMiddleSpaces and added skipTrailingSpaces
hasMiddleSpaces was conflicting with the new function (skipTrailingSpaces) I was attempting to create, so I removed it and incorporated most of its functionality within skipTrailingSpaces. The new function checks if there are any letters after a space, and counts it as false if there is. I also reworked some of validString because it was not working 100% right.